Use shutil.copy for file copy
This change was made for performance reasons as copying should be faster than opening both files and writing contents. The thing is the dataset could be huge and could take a while even for copying. There's no known way to copy a file with a progress bar without overcomplicating things, so I've used a simple print before and after to indicate when the copying happens and finishes.